+++ src/components/tabs/tabs.tsx
@@ -10,8 +10,8 @@ export const TabsList = withClassName(
 
 export const TabsTrigger = withClassName(
 	Tabs.Trigger,
-	'flex flex-row items-center justify-center gap-2 color-black py-1 px-4 bg-wash text-md font-bold min-w-100px rounded-full border-none transition-colors cursor-pointer select-none',
-	'hover:bg-primary-light focus-visible:(shadow-focus outline-off) [&[data-state=active]]:(bg-primary-dark text-white cursor-default hover:bg-primary-dark relative z-1)',
+	'flex flex-row items-center justify-center gap-2 color-black py-1 px-4 bg-wash text-md min-w-100px rounded-full border border-1 border-primary border-solid transition-colors cursor-pointer select-none',
+	'hover:bg-primary-light focus-visible:(focus-ring focus-ring-primary-dark outline-off bg-primary-light border-primary-dark) [&[data-state=active]]:(font-bold bg-primary-light border-primary-light text-black cursor-default hover:bg-primary-light relative z-1)',
 );
 
 export const TabsContent = withClassName(Tabs.Content, '');